refactor (index.ts)
