+++ code/system/sys.data.crdt/index.html
@@ -5,6 +5,6 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   </head>
   <body>
-    <script type="module" src="src/sample/entry.mts"></script>
+    <script type="module" src="src/sample/entry.ts"></script>
   </body>
 </html>
