+++ code/system/sys.data.crdt/src/test/index.ts
@@ -3,4 +3,4 @@ export { describe, it } from 'vitest';
 export { expect, expectError } from 'sys.test';
 export { TestFilesystem } from 'sys.fs';
 
-export * from '../common/index.mjs';
+export * from '../common/';
