+++ code/system/sys.text/src/Markdown.Processor/types/common.ts
