+++ code/system/sys.text/src/Markdown.Processor/types/types.code.mts
@@ -1,4 +1,4 @@
-import { t } from './common.mjs';
+import { t } from './common';
 
 /**
  * Represents a code-block within Markdown.
