+++ code/system/sys.text/src/Markdown/Markdown.Find.mts
@@ -1,5 +1,5 @@
 import { CONTINUE, EXIT, visit } from 'unist-util-visit';
-import { t } from '../common.mjs';
+import { t } from '../common';
 
 /**
  * Flags for looking at markdown.
