+++ code/system/sys.text/src/Markdown/Markdown.mts
@@ -1,6 +1,6 @@
-import { t } from '../common.mjs';
+import { t } from '../common';
 import { MarkdownIs as Is } from './Markdown.Is.mjs';
-import { MarkdownProcessor as processor } from '../Markdown.Processor/index.mjs';
+import { MarkdownProcessor as processor } from '../Markdown.Processor';
 import { MarkdownFind as Find } from './Markdown.Find.mjs';
 
 export const Markdown: t.Markdown = {
