+++ code/system/sys.text/src/Text.Processor/index.ts
