+++ code/system/sys.text/src/Text/index.ts
