--- code/system/sys.text/src/common.mts
@@ -1,1 +1,0 @@
-export * from './common/index.mjs';
