+++ code/system/sys.text/src/common/index.ts
