+++ code/system/sys.text/src/index.mts
@@ -1,3 +1,3 @@
 export { Pkg } from './index.pkg.mjs';
-export { Text } from './Text/index.mjs';
+export { Text } from './Text';
 export { Markdown } from './Markdown';
