+++ code/system/sys.text/src/test/index.ts
@@ -1,4 +1,4 @@
 export { expect, expectError } from 'sys.test';
 export { describe, it } from 'vitest';
 
-export * from '../common.mjs';
+export * from '../common';
